+++ tee-worker/omni-executor/client-sdk/packages/client-sdk/src/lib/requests/__test__/pumpx-add-wallet.request.test.ts
@@ -9,7 +9,9 @@ describe.skip('add-export-wallet', () => {
     const rpcRequest = {
       jsonrpc: '2.0',
       method: 'omni_requestEmailVerificationCode',
-      params: ['test@google.com'],
+      params: {
+        user_email: 'test@google.com',
+      },
     };
     await enclave.send(rpcRequest);
   });
@@ -19,43 +21,32 @@ describe.skip('add-export-wallet', () => {
     const rpcRequest: JsonRpcRequest = {
       jsonrpc: '2.0',
       method: 'pumpx_requestJwt',
-      params: ['test@google.com', '', '', '', '1223'],
+      params: {
+        user_email: 'test@google.com',
+        invite_code: '', // Optional
+        google_code: '', // Optional
+        language: 'en', // Optional
+        email_code: '98121',
+      },
     };
 
     const response = await enclave.send(rpcRequest);
     // console.log(response.auth_token);
   });
+
+  // step 3
   it('add wallet should works', async () => {
     const rpcRequest: JsonRpcRequest = {
       jsonrpc: '2.0',
       method: PumpxRpcMethod.AddWallet,
       id: 1,
-      params: [
-        'test@google.com', // user_email
-        '0xxxxxx', // auth_token already get from step 2
-      ],
+      params: {
+        user_email: 'test@google.com',
+        auth_token:
+          '"eyJOeXAi0iJKV1QiLCJhbGci0iJSUzI1NiJ9.ey...6FcLY4XByKuT3a5Ax3L0En4LFbOLdGwrsNZDMOEU...iJhFisr4KHFFdjhphK_G9qn45vQ"',  //response.auth_token from step 2
+      },
     };
 
-    await enclave.send(rpcRequest);
-  });
-
-  it('export wallet should works', async () => {
-    const aesRandomKey = globalThis.crypto.getRandomValues(new Uint8Array(32));
-    const encryptedKey = await enclave.encrypt({ cleartext: aesRandomKey }); // already get shielding key
-    const rpcRequest: JsonRpcRequest = {
-      jsonrpc: '2.0',
-      method: PumpxRpcMethod.ExportWallet,
-      id: 1,
-      params: [
-        'test@google.com', // user_email
-        encryptedKey.ciphertext.toString(), // key
-        '123', // google_code
-        '0', // chain_id
-        '1', // wallet_index
-        '0xdeadbeaf', // wallet_address
-        '123', // email_code
-      ],
-    };
     await enclave.send(rpcRequest);
   });
 });
